refactor code shared between controllers
