+++ src/main/java/application/MenuController.java
@@ -35,30 +35,19 @@ import java.util.Arrays;
 import java.util.Date;
 
 /**
- * Created by willpride on 1/12/16.
+ * Controller (API endpoint) containing all session navigation functionality.
+ * This includes module, form, case, and session (incomplete form) selection.
  */
 @Api(value = "Menu Controllers", description = "Operations for navigating CommCare Menus and Cases")
 @RestController
 @EnableAutoConfiguration
-public class MenuController {
+public class MenuController extends AbstractBaseController{
 
     @Value("${commcarehq.host}")
     private String host;
 
     @Autowired
-    private RestoreService restoreService;
-
-    @Autowired
     private InstallService installService;
-
-    @Autowired
-    private SessionRepo sessionRepo;
-
-    @Autowired
-    private JavaMailSenderImpl exceptionSender;
-
-    @Autowired
-    private SimpleMailMessage exceptionMessage;
 
     private final Log log = LogFactory.getLog(MenuController.class);
 
@@ -76,7 +65,7 @@ public class MenuController {
      * Make a a series of menu selections (as above, but can have multiple)
      *
      * @param sessionNavigationBean Give an installation code or path and a set of session selections
-     * @param authToken The Django session id auth token
+     * @param authToken             The Django session id auth token
      * @return A MenuBean or a NewFormSessionResponse
      * @throws Exception
      */
@@ -89,14 +78,14 @@ public class MenuController {
         MenuSession menuSession = performInstall(sessionNavigationBean, authToken);
         String[] selections = sessionNavigationBean.getSelections();
         Object nextMenu = getNextMenu(menuSession);
-        if (selections == null){
+        if (selections == null) {
             log.info("Selections null, got next menu: " + nextMenu);
             System.out.println("Menu Session Options: " + Arrays.toString(menuSession.getNextScreen().getOptions()));
             return nextMenu;
         }
-        for(String selection: selections) {
+        for (String selection : selections) {
             menuSession.handleInput(selection);
-            if(menuSession.getNextScreen() != null){
+            if (menuSession.getNextScreen() != null) {
                 System.out.println("Menu Session Options: " + Arrays.toString(menuSession.getNextScreen().getOptions()));
             }
         }
@@ -106,17 +95,17 @@ public class MenuController {
     }
 
     private MenuSession performInstall(InstallRequestBean bean, String authToken) throws Exception {
-        if((bean.getAppId() == null || "".equals(bean.getAppId())) &&
-                bean.getInstallReference() == null || "".equals(bean.getInstallReference())){
+        if ((bean.getAppId() == null || "".equals(bean.getAppId())) &&
+                bean.getInstallReference() == null || "".equals(bean.getInstallReference())) {
             throw new RuntimeException("Either app_id or install_reference must be non-null.");
         }
 
         HqAuth auth;
-        if(authToken != null && !authToken.trim().equals("")){
+        if (authToken != null && !authToken.trim().equals("")) {
             auth = new DjangoAuth(authToken);
-        } else{
+        } else {
             String password = bean.getPassword();
-            if(password == null || "".equals(password.trim())){
+            if (password == null || "".equals(password.trim())) {
                 throw new RuntimeException("Either authToken or password must be non-null");
             }
             auth = new BasicAuth(bean.getUsername(), bean.getDomain(), "commcarehq.org", password);
@@ -146,31 +135,30 @@ public class MenuController {
         nextScreen = menuSession.getNextScreen();
 
         // No next menu screen? Start form entry!
-        if (nextScreen == null){
+        if (nextScreen == null) {
             return generateFormEntryScreen(menuSession);
-        }
-        else{
+        } else {
             MenuBean menuResponseBean;
 
             // We're looking at a module or form menu
-            if(nextScreen instanceof MenuScreen){
+            if (nextScreen instanceof MenuScreen) {
                 menuResponseBean = generateMenuScreen((MenuScreen) nextScreen, menuSession.getSessionWrapper());
             }
             // We're looking at a case list or detail screen (probably)
             else if (nextScreen instanceof EntityScreen) {
                 menuResponseBean = generateEntityScreen((EntityScreen) nextScreen, offset, searchText);
-            } else{
+            } else {
                 throw new Exception("Unable to recognize next screen: " + nextScreen);
             }
             return menuResponseBean;
         }
     }
 
-    private CommandListResponseBean generateMenuScreen(MenuScreen nextScreen, SessionWrapper session){
+    private CommandListResponseBean generateMenuScreen(MenuScreen nextScreen, SessionWrapper session) {
         return new CommandListResponseBean(nextScreen, session);
     }
 
-    private EntityListResponse generateEntityScreen(EntityScreen nextScreen, int offset, String searchText){
+    private EntityListResponse generateEntityScreen(EntityScreen nextScreen, int offset, String searchText) {
         return new EntityListResponse(nextScreen, offset, searchText);
     }
 
@@ -179,37 +167,5 @@ public class MenuController {
         sessionRepo.save(formEntrySession.serialize());
         log.info("Start form entry with session: " + formEntrySession);
         return new NewFormSessionResponse(formEntrySession);
-    }
-
-    @ExceptionHandler(Exception.class)
-    public String handleError(HttpServletRequest req, Exception exception) {
-        log.error("Request: " + req.getRequestURL() + " raised " + exception);
-        exception.printStackTrace();
-        sendExceptionEmail(exception);
-        JSONObject errorReturn = new JSONObject();
-        errorReturn.put("exception", exception);
-        errorReturn.put("url", req.getRequestURL());
-        errorReturn.put("status", "error");
-        return errorReturn.toString();
-    }
-
-    private void sendExceptionEmail(Exception exception) {
-        exceptionMessage.setText(getExceptionEmailBody(exception));
-        exceptionMessage.setSubject("Formplayer Menu Exception: " + exception.getMessage());
-        try {
-            exceptionSender.send(exceptionMessage);
-        } catch(MailSendException e){
-            // I think we should fail quietly on this
-            log.error("Couldn't send exception email: " + e);
-        }
-    }
-
-
-    private String getExceptionEmailBody(Exception exception){
-        DateFormat dateFormat = new SimpleDateFormat("yyyy/MM/dd HH:mm:ss");
-        String formattedTime = dateFormat.format(new Date());
-        return "Message: " + exception.getMessage() + " \n " +
-                "Time : " + formattedTime + " \n " +
-                "Trace: " + Arrays.toString(exception.getStackTrace());
     }
 }
